+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\roelo\\WebstormProjects\\frontend-react-pokemon\\src\\index.js":"1","C:\\Users\\roelo\\WebstormProjects\\frontend-react-pokemon\\src\\App.js":"2","C:\\Users\\roelo\\WebstormProjects\\frontend-react-pokemon\\src\\reportWebVitals.js":"3","C:\\Users\\roelo\\WebstormProjects\\frontend-react-pokemon\\src\\components\\PokemonCard.js":"4"},{"size":494,"mtime":1633009554385,"results":"5","hashOfConfig":"6"},{"size":1598,"mtime":1633375492880,"results":"7","hashOfConfig":"6"},{"size":375,"mtime":1633009554386,"results":"8","hashOfConfig":"6"},{"size":644,"mtime":1633375384840,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},"mrhc85",{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\roelo\\WebstormProjects\\frontend-react-pokemon\\src\\index.js",[],["19","20"],"C:\\Users\\roelo\\WebstormProjects\\frontend-react-pokemon\\src\\App.js",[],"C:\\Users\\roelo\\WebstormProjects\\frontend-react-pokemon\\src\\reportWebVitals.js",[],"C:\\Users\\roelo\\WebstormProjects\\frontend-react-pokemon\\src\\components\\PokemonCard.js",[],{"ruleId":"21","replacedBy":"22"},{"ruleId":"23","replacedBy":"24"},"no-native-reassign",["25"],"no-negated-in-lhs",["26"],"no-global-assign","no-unsafe-negation"]
+[{"C:\\Users\\roelo\\WebstormProjects\\frontend-react-pokemon\\src\\index.js":"1","C:\\Users\\roelo\\WebstormProjects\\frontend-react-pokemon\\src\\App.js":"2","C:\\Users\\roelo\\WebstormProjects\\frontend-react-pokemon\\src\\reportWebVitals.js":"3","C:\\Users\\roelo\\WebstormProjects\\frontend-react-pokemon\\src\\components\\PokemonCard.js":"4"},{"size":494,"mtime":1633009554385,"results":"5","hashOfConfig":"6"},{"size":1858,"mtime":1633420638241,"results":"7","hashOfConfig":"6"},{"size":375,"mtime":1633009554386,"results":"8","hashOfConfig":"6"},{"size":1785,"mtime":1633419985100,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},"mrhc85",{"filePath":"13","messages":"14","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\roelo\\WebstormProjects\\frontend-react-pokemon\\src\\index.js",[],["19","20"],"C:\\Users\\roelo\\WebstormProjects\\frontend-react-pokemon\\src\\App.js",["21","22"],"C:\\Users\\roelo\\WebstormProjects\\frontend-react-pokemon\\src\\reportWebVitals.js",[],"C:\\Users\\roelo\\WebstormProjects\\frontend-react-pokemon\\src\\components\\PokemonCard.js",[],{"ruleId":"23","replacedBy":"24"},{"ruleId":"25","replacedBy":"26"},{"ruleId":"27","severity":1,"message":"28","line":1,"column":26,"nodeType":"29","messageId":"30","endLine":1,"endColumn":35},{"ruleId":"27","severity":1,"message":"31","line":2,"column":8,"nodeType":"29","messageId":"30","endLine":2,"endColumn":13},"no-native-reassign",["32"],"no-negated-in-lhs",["33"],"no-unused-vars","'useEffect' is defined but never used.","Identifier","unusedVar","'axios' is defined but never used.","no-global-assign","no-unsafe-negation"]
